+++ src/RecoveryServices/RecoveryServices.Backup.Helpers/Conversions/RetentionPolicyConversions.cs
@@ -12,12 +12,11 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models;
+using Microsoft.Azure.Commands.RecoveryServices.Backup.Properties;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.Models;
-using Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers;
-using Microsoft.Azure.Commands.RecoveryServices.Backup.Properties;
 using ServiceClientModel = Microsoft.Azure.Management.RecoveryServices.Backup.Models;
 
 namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
@@ -75,7 +74,7 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
         }
 
         public static SimpleRetentionPolicy GetPSSimpleRetentionPolicy(
-           ServiceClientModel.SimpleRetentionPolicy hydraRetPolicy, string timeZone)
+           ServiceClientModel.SimpleRetentionPolicy hydraRetPolicy, string timeZone, string provider)
         {
             if (hydraRetPolicy == null)
             {
@@ -92,7 +91,24 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
                     (int)hydraRetPolicy.RetentionDuration.Count : default(int);
             }
 
-            simplePolicy.Validate();
+            if (string.Compare(provider, "AzureSql") == 0)
+            {
+                int weeklyLimit = PolicyConstants.MaxAllowedRetentionDurationCountWeeklySql;
+                int monthlyLimit = PolicyConstants.MaxAllowedRetentionDurationCountMonthlySql;
+                int yearlyLimit = PolicyConstants.MaxAllowedRetentionDurationCountYearlySql;
+
+                if ((simplePolicy.RetentionDurationType == RetentionDurationType.Days) ||
+                    (simplePolicy.RetentionDurationType == RetentionDurationType.Weeks &&
+                        (simplePolicy.RetentionCount <= 0 || simplePolicy.RetentionCount > weeklyLimit)) ||
+                    (simplePolicy.RetentionDurationType == RetentionDurationType.Months &&
+                        (simplePolicy.RetentionCount <= 0 || simplePolicy.RetentionCount > monthlyLimit)) ||
+                    (simplePolicy.RetentionDurationType == RetentionDurationType.Years &&
+                        (simplePolicy.RetentionCount <= 0 || simplePolicy.RetentionCount > yearlyLimit)))
+                {
+                    throw new ArgumentException(Resources.AllowedSqlRetentionRange);
+                }
+            }
+
             return simplePolicy;
         }