+++ src/Automation/Automation.Test/UnitTests/ImportAzureAutomationDscNodeConfigurationTest.cs
@@ -16,13 +16,12 @@ using Microsoft.Azure.Commands.Automation.Cmdlet;
 using Microsoft.Azure.Commands.Automation.Common;
 using Microsoft.WindowsAzure.Commands.Common.Test.Mocks;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
-using Microsoft.WindowsAzure.Commands.Test.Utilities.Common;
 using Moq;
 using Xunit;
 
 namespace Microsoft.Azure.Commands.ResourceManager.Automation.Test.UnitTests
 {
-    public class ImportAzureAutomationDscNodeConfigurationTest : RMTestBase
+    public class ImportAzureAutomationDscNodeConfigurationTest
     {
         private Mock<IAutomationPSClient> mockAutomationClient;