+++ src/Compute/Compute.Test/ScenarioTests/ImageTests.ps1
@@ -125,15 +125,54 @@ function Test-Image
         Assert-AreEqual "testval2" $createdImage.Tags["test2"]
 
         # List and Delete Image
+        $wildcardRgQuery = ($rgname -replace ".$") + "*"
+        $wildcardNameQuery = ($imageName -replace ".$") + "*"
+        
+        $images = Get-AzImage;
+        Assert-True { $images.Count -ge 1 };
+        
+        $images = Get-AzImage -ResourceGroupName $wildcardRgQuery;
+        Assert-AreEqual 1 $images.Count;
+        Assert-AreEqual $rgname $images[0].ResourceGroupName;
+
         $images = Get-AzImage -ResourceGroupName $rgname;
         Assert-AreEqual 1 $images.Count;
+        Assert-AreEqual $rgname $images[0].ResourceGroupName;
+        
+        $images = Get-AzImage -Name $wildcardNameQuery;
+        Assert-AreEqual 1 $images.Count;
+        Assert-AreEqual $rgname $images[0].ResourceGroupName;
+        Assert-AreEqual $imageName $images[0].Name;
+        
+        $images = Get-AzImage -Name $imageName;
+        Assert-AreEqual 1 $images.Count;
+        Assert-AreEqual $rgname $images[0].ResourceGroupName;
+        Assert-AreEqual $imageName $images[0].Name;
+        
+        $images = Get-AzImage -ResourceGroupName $wildcardRgQuery -Name $wildcardNameQuery;
+        Assert-AreEqual 1 $images.Count;
+        Assert-AreEqual $rgname $images[0].ResourceGroupName;
+        Assert-AreEqual $imageName $images[0].Name;
+        
+        $images = Get-AzImage -ResourceGroupName $rgname -Name $wildcardNameQuery;
+        Assert-AreEqual 1 $images.Count;
+        Assert-AreEqual $rgname $images[0].ResourceGroupName;
+        Assert-AreEqual $imageName $images[0].Name;
+        
+        $images = Get-AzImage -ResourceGroupName $wildcardRgQuery -Name $imageName;
+        Assert-AreEqual 1 $images.Count;
+        Assert-AreEqual $rgname $images[0].ResourceGroupName;
+        Assert-AreEqual $imageName $images[0].Name;
+        
+        $images = Get-AzImage -ResourceGroupName $rgname -Name $imageName;
+        Assert-AreEqual $rgname $images.ResourceGroupName;
+        Assert-AreEqual $imageName $images.Name;
 
         # Update Image Tag
-        $image = Get-AzImage -ResourceGroupName $rgname -ImageName $imageName;
-        $image.Tags =  New-Object "System.Collections.Generic.Dictionary``2[System.String,System.String]";
-        $image.Tags.Add("test1", "testval3");
-        $image.Tags.Add("test2", "testval4");
-        Update-AzImage -ResourceGroupName $rgname -ImageName $imageName -Image $image;
+        $images[0] | Update-AzImage -Tag @{test1 = "testval3"; test2 = "testval4"};
+        Update-AzImage -ResourceGroupName $rgname -ImageName $imageName -Tag @{test1 = "testval3"; test2 = "testval4"};
+        Update-AzImage -Image $images[0] -Tag @{test1 = "testval3"; test2 = "testval4"};
+        Update-AzImage -ResourceId $images[0].Id -Tag @{test1 = "testval3"; test2 = "testval4"};
 
         $image = Get-AzImage -ResourceGroupName $rgname -ImageName $imageName;
         Assert-True {$image.Tags.ContainsKey("test1") }
@@ -215,7 +254,7 @@ function Test-ImageCapture
         $p = Add-AzVMDataDisk -VM $p -Name 'testDataDisk2' -Caching 'ReadOnly' -DiskSizeInGB 11 -Lun 2 -VhdUri $dataDiskVhdUri2 -CreateOption Empty;
         $p = Add-AzVMDataDisk -VM $p -Name 'testDataDisk3' -Caching 'ReadOnly' -DiskSizeInGB 12 -Lun 3 -VhdUri $dataDiskVhdUri3 -CreateOption Empty;
         $p = Remove-AzVMDataDisk -VM $p -Name 'testDataDisk3';
-        
+
         # OS & Image
         $user = "Foo12";
         $password = $PLACEHOLDER;