+++ src/Sql/Sql/Auditing/Cmdlet/SetAzSqlServerAuditing.cs
@@ -16,14 +16,15 @@ using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Commands.Sql.Auditing.Model;
 using Microsoft.Azure.Commands.Sql.Common;
 using Microsoft.Azure.Commands.Sql.Server.Model;
+using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using System;
 using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Sql.Auditing.Cmdlet
 {
-    /// <summary>
-    /// Sets the auditing settings properties for a specific database server.
-    /// </summary>
+    [CmdletOutputBreakingChange(
+        typeof(DatabaseBlobAuditingSettingsModel),
+        ReplacementCmdletOutputTypeName = "bool")]
     [Cmdlet(
         VerbsCommon.Set,
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "SqlServerAuditing",
@@ -92,22 +93,22 @@ namespace Microsoft.Azure.Commands.Sql.Auditing.Cmdlet
             ParameterSetName = DefinitionsCommon.BlobStorageByParentResourceParameterSetName,
             Mandatory = true,
             ValueFromPipeline = true,
-            HelpMessage = AuditingHelpMessages.DatabaseInputObjectHelpMessage)]
+            HelpMessage = AuditingHelpMessages.ServerInputObjectHelpMessage)]
         [Parameter(
             ParameterSetName = DefinitionsCommon.StorageAccountSubscriptionIdByParentResourceParameterSetName,
             Mandatory = true,
             ValueFromPipeline = true,
-            HelpMessage = AuditingHelpMessages.DatabaseInputObjectHelpMessage)]
+            HelpMessage = AuditingHelpMessages.ServerInputObjectHelpMessage)]
         [Parameter(
             ParameterSetName = DefinitionsCommon.EventHubByParentResourceParameterSetName,
             Mandatory = true,
             ValueFromPipeline = true,
-            HelpMessage = AuditingHelpMessages.DatabaseInputObjectHelpMessage)]
+            HelpMessage = AuditingHelpMessages.ServerInputObjectHelpMessage)]
         [Parameter(
             ParameterSetName = DefinitionsCommon.LogAnalyticsByParentResourceParameterSetName,
             Mandatory = true,
             ValueFromPipeline = true,
-            HelpMessage = AuditingHelpMessages.DatabaseInputObjectHelpMessage)]
+            HelpMessage = AuditingHelpMessages.ServerInputObjectHelpMessage)]
         [ValidateNotNullOrEmpty]
         public override AzureSqlServerModel InputObject { get; set; }