+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -20,6 +20,7 @@
 ## Upcoming Release
 * Added Sql server in Azure VM support
 * SDK Update
+* Removed VMappContainer check in Get-ProtectableItem
 
 ## Version 1.0.1
 * Release with updated Authentication dependency