+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added Sql server in Azure VM support
+* SDK Update
+
+## Version 1.0.1
+* Release with updated Authentication dependency
 
 ## Version 1.0.0
 * General availability of `Az.RecoveryServices` module